Day 2
We went over everything that we did yesterday. we got into event handlers and what they are we also used a couple of them. we mainly used the onclick event handler and we were stoped early by a fire drill when we were creating the button that would switch the puzzle.